+++ fuc/api/pymaf.py
@@ -240,10 +240,11 @@ def legend_handles(name='regular'):
     return handles
 
 class AnnFrame:
-    """Class for storing sample annotation data.
+    """
+    Class for storing sample annotation data.
 
-    This class stores sample annotation data as pandas.DataFrame with sample
-    names as index.
+    This class stores sample annotation data as :class:`pandas.DataFrame`
+    with sample names as index.
 
     Note that an AnnFrame can have a different set of samples than its
     accompanying MafFrame.
@@ -283,14 +284,15 @@ class AnnFrame:
     Sara_N                  Sara  Normal   57
     Sara_T                  Sara   Tumor   57
     """
-    def __init__(self, df):
-        self._df = self._check_df(df)
 
     def _check_df(self, df):
         if type(df.index) == pd.RangeIndex:
             m = "Index must be sample names, not 'pandas.RangeIndex'."
             raise ValueError(m)
         return df
+
+    def __init__(self, df):
+        self._df = self._check_df(df)
 
     @property
     def df(self):
@@ -300,6 +302,33 @@ class AnnFrame:
     @df.setter
     def df(self, value):
         self._df = self._check_df(value)
+
+    @property
+    def samples(self):
+        """list : List of the sample names."""
+        return list(self.df.index.to_list())
+
+    @property
+    def shape(self):
+        """tuple : Dimensionality of AnnFrame (samples, annotations)."""
+        return self.df.shape
+
+    def filter_mf(self, mf):
+        """
+        Filter the AnnFrame for the samples in the MafFrame.
+
+        Parameters
+        ----------
+        mf : MafFrame
+            MafFrame containing target samples.
+
+        Returns
+        -------
+        AnnFrame
+            Filtered AnnFrame object.
+        """
+        df = self.df.loc[mf.samples]
+        return self.__class__(df)
 
     @classmethod
     def from_dict(cls, data, sample_col='Tumor_Sample_Barcode'):
@@ -352,7 +381,8 @@ class AnnFrame:
 
     @classmethod
     def from_file(cls, fn, sample_col='Tumor_Sample_Barcode', sep='\t'):
-        """Construct AnnFrame from a delimited text file.
+        """
+        Construct an AnnFrame from a delimited text file.
 
         The text file must have at least one column that represents
         sample names which are used as index for pandas.DataFrame.
@@ -491,11 +521,15 @@ class AnnFrame:
         self, col, samples=None, numeric=False, segments=5, decimals=0,
         cmap='Pastel1', ax=None, figsize=None, **kwargs
     ):
-        """Create a 1D categorical heatmap for the given column.
+        """
+        Create a 1D categorical heatmap of the column.
 
-        In the case of a numeric column, use ``numeric=True`` which will
+        In the case of a numeric column, set ``numeric`` as True which will
         divide the values into equal-sized intervals, with the number of
-        intervals determined by the `segments` option.
+        intervals determined by ``segments``.
+
+        See the :ref:`tutorials:Create customized oncoplots` tutorial to
+        learn how to create customized oncoplots.
 
         Parameters
         ----------
@@ -559,6 +593,25 @@ class AnnFrame:
         ax.set_yticks([])
         return ax
 
+    def sorted_samples(self, by, mf=None, keep_empty=False, nonsyn=False):
+        """
+        Return a sorted list of sample names.
+
+        Parameters
+        ----------
+        df : str or list
+            Column or list of columns to sort by.
+        """
+        df = self.df.copy()
+
+        if nonsyn:
+            samples = mf.matrix_waterfall(keep_empty=keep_empty).columns
+            df = df.loc[samples]
+
+        df = df.sort_values(by=by)
+
+        return df.index.to_list()
+
 class MafFrame:
     """Class for storing MAF data.
 
@@ -592,22 +645,31 @@ class MafFrame:
 
     @classmethod
     def from_file(cls, fn):
-        """Construct MafFrame from a MAF file.
+        """
+        Construct a MafFrame from a MAF file.
 
         Parameters
         ----------
         fn : str
-            MAF file path (zipped or unzipped).
+            MAF file (zipped or unzipped).
 
         Returns
         -------
         MafFrame
-            MafFrame.
+            MafFrame object.
 
         See Also
         --------
         MafFrame
             MafFrame object creation using constructor.
+
+        Examples
+        --------
+
+        >>> from fuc import common, pymaf
+        >>> common.load_dataset('tcga-laml')
+        >>> fn = '~/fuc-data/tcga-laml/tcga_laml.maf.gz'
+        >>> mf = pymaf.MafFrame.from_file(fn)
         """
         # Read the MAF file.
         df = pd.read_table(fn)
@@ -631,11 +693,14 @@ class MafFrame:
         if case_dict:
             df = df.rename(columns=case_dict)
 
+        # Set the data types.
+        df.Chromosome = df.Chromosome.astype(str)
+
         return cls(df)
 
     @classmethod
-    def from_vcf(cls, vcf):
-        """Construct MafFrame from a VCF file or VcfFrame.
+    def from_vcf(cls, vcf, keys=None, names=None):
+        """Construct a MafFrame from a VCF file or VcfFrame.
 
         The input VCF should already contain functional annotation data
         from a tool such as Ensemble VEP, SnpEff, and ANNOVAR. The
@@ -647,6 +712,10 @@ class MafFrame:
         ----------
         vcf : str or VcfFrame
             VCF file or VcfFrame.
+        keys : str or list
+            Genotype key or list of genotype keys.
+        names : str or list
+            Column name or list of column names to use in the MafFrame.
 
         Examples
         --------
@@ -748,6 +817,10 @@ class MafFrame:
                 Tumor_Seq_Allele2 = tumor_seq_allele2,
                 Tumor_Sample_Barcode = tumor_sample_barcode,
                 Protein_Change = protein_change,
+                CHROM = r.CHROM,
+                POS = r.POS,
+                REF = r.REF,
+                ALT = r.ALT,
             )
 
             return pd.Series(d)
@@ -763,10 +836,38 @@ class MafFrame:
         del df['Tumor_Sample_Barcode']
         df = df.join(s)
 
+        # Append the genotype keys.
+        if keys is None:
+            keys = []
+        if names is None:
+            names = []
+        if isinstance(keys, str):
+            keys = [keys]
+        if isinstance(names, str):
+            names = [names]
+        for i, key in enumerate(keys):
+            temp_df = vf.extract(key)
+            temp_df = pd.concat([vf.df.iloc[:, :9], temp_df], axis=1)
+            temp_df = temp_df.drop(
+                columns=['ID', 'QUAL', 'FILTER', 'INFO', 'FORMAT'])
+            temp_df = pd.melt(
+                temp_df,
+                id_vars=['CHROM', 'POS', 'REF', 'ALT'],
+                var_name='Tumor_Sample_Barcode',
+            )
+            temp_df = temp_df[temp_df.value != '.']
+            df = df.merge(temp_df,
+                on=['CHROM', 'POS', 'REF', 'ALT', 'Tumor_Sample_Barcode'])
+            df = df.rename(columns={'value': names[i]})
+
+        df = df.drop(columns=['CHROM', 'POS', 'REF', 'ALT'])
+
         return cls(df)
 
-    def compute_genes(self, count=10, mode='variants'):
-        """Compute a matrix of counts for genes and variant classifications.
+    def matrix_genes(self, count=10, mode='variants'):
+        """
+        Compute a matrix of variant counts with a shape of (genes, variant
+        classifications).
 
         Parameters
         ----------
@@ -796,7 +897,7 @@ class MafFrame:
             df = df[:count]
             df = df.rename_axis(None, axis=1)
         elif mode == 'samples':
-            df = self.compute_waterfall(count)
+            df = self.matrix_waterfall(count)
             df = df.apply(lambda r: r.value_counts(), axis=1)
             for varcls in NONSYN_NAMES + ['Multi_Hit']:
                 if varcls not in df.columns:
@@ -807,8 +908,10 @@ class MafFrame:
             raise ValueError(f'Found incorrect mode: {mode}')
         return df
 
-    def compute_tmb(self):
-        """Compute a matrix of counts for samples and variant classifications.
+    def matrix_tmb(self):
+        """
+        Compute a matrix of variant counts with a shape of (samples, variant
+        classifications).
 
         Returns
         -------
@@ -832,13 +935,20 @@ class MafFrame:
         df = df.rename_axis(None, axis=1)
         return df
 
-    def compute_waterfall(self, count=10):
-        """Compute a matrix of variant classifications for genes and samples.
+    def matrix_waterfall(self, count=10, keep_empty=False):
+        """
+        Compute a matrix of variant classifications with a shape of
+        (genes, samples).
+
+        If there are multiple variant classifications available for a given
+        cell, they will be replaced as 'Multi_Hit'.
 
         Parameters
         ----------
         count : int, default: 10
-            Number of top mutated genes to display.
+            Number of top mutated genes to include.
+        keep_empty : bool, default: False
+            If True, keep samples with all ``NaN``'s.
 
         Returns
         -------
@@ -846,6 +956,7 @@ class MafFrame:
             Dataframe containing waterfall data.
         """
         df = self.df[self.df.Variant_Classification.isin(NONSYN_NAMES)]
+
         f = lambda x: ''.join(x) if len(x) == 1 else 'Multi_Hit'
         df = df.groupby(['Hugo_Symbol', 'Tumor_Sample_Barcode'])[
             'Variant_Classification'].apply(f).to_frame()
@@ -860,8 +971,9 @@ class MafFrame:
         # Select the top mutated genes.
         df = df[:count]
 
-        # Remove columns (samples) with all NaN's.
-        df = df.dropna(axis=1, how='all')
+        # Drop samples with all NaN's.
+        if not keep_empty:
+            df = df.dropna(axis=1, how='all')
 
         # Sort the columns (samples).
         c = df.applymap(lambda x: 0 if pd.isnull(x) else 1).sort_values(
@@ -921,7 +1033,7 @@ class MafFrame:
             colors = NONSYN_COLORS + ['k']
         else:
             raise ValueError(f'Found incorrect mode: {mode}')
-        df = self.compute_genes(count, mode=mode)
+        df = self.matrix_genes(count, mode=mode)
         df = df.iloc[::-1]
         if ax is None:
             fig, ax = plt.subplots(figsize=figsize)
@@ -932,15 +1044,21 @@ class MafFrame:
         return ax
 
     def plot_oncoplot(
-        self, count=10, figsize=(15, 10), label_fontsize=15,
+        self, count=10, keep_empty=False, figsize=(15, 10), label_fontsize=15,
         ticklabels_fontsize=15, legend_fontsize=15
     ):
-        """Create an oncoplot.
+        """
+        Create a standard oncoplot.
+
+        See the :ref:`tutorials:Create customized oncoplots` tutorial to
+        learn how to create customized oncoplots.
 
         Parameters
         ----------
         count : int, default: 10
             Number of top mutated genes to display.
+        keep_empty : bool, default: False
+            If True, display samples that do not have any mutations.
         figsize : tuple, default: (15, 10)
             Width, height in inches. Format: (float, float).
         label_fontsize : float, default: 15
@@ -967,14 +1085,14 @@ class MafFrame:
         [[ax1, ax2], [ax3, ax4], [ax5, ax6]] = axes
 
         # Create the TMB plot.
-        samples = list(self.compute_waterfall(count).columns)
+        samples = list(self.matrix_waterfall(count=count, keep_empty=keep_empty).columns)
         self.plot_tmb(ax=ax1, samples=samples)
         ax1.set_xlabel('')
         ax1.spines['right'].set_visible(False)
         ax1.spines['top'].set_visible(False)
         ax1.spines['bottom'].set_visible(False)
         ax1.set_ylabel('TMB', fontsize=label_fontsize)
-        ax1.set_yticks([0, self.compute_tmb().sum(axis=1).max()])
+        ax1.set_yticks([0, self.matrix_tmb().sum(axis=1).max()])
         ax1.tick_params(axis='y', which='major',
                         labelsize=ticklabels_fontsize)
 
@@ -982,7 +1100,7 @@ class MafFrame:
         ax2.remove()
 
         # Create the waterfall plot.
-        self.plot_waterfall(count=count, ax=ax3, linewidths=1)
+        self.plot_waterfall(count=count, ax=ax3, linewidths=1, keep_empty=keep_empty)
         ax3.set_xlabel('')
         ax3.tick_params(axis='y', which='major', labelrotation=0,
                         labelsize=ticklabels_fontsize)
@@ -994,7 +1112,7 @@ class MafFrame:
         ax4.spines['top'].set_visible(False)
         ax4.set_yticks([])
         ax4.set_xlabel('Samples', fontsize=label_fontsize)
-        ax4.set_xticks([0, self.compute_genes(
+        ax4.set_xticks([0, self.matrix_genes(
             10, mode='samples').sum(axis=1).max()])
         ax4.set_ylim(-0.5, count-0.5)
         ax4.tick_params(axis='x', which='major',
@@ -1262,7 +1380,7 @@ class MafFrame:
                         labelsize=ticklabels_fontsize)
 
         # Create the 'Variants per sample' figure.
-        median = self.compute_tmb().sum(axis=1).median()
+        median = self.matrix_tmb().sum(axis=1).median()
         self.plot_tmb(ax=ax4)
         ax4.set_title(f'Variants per sample (median={median:.1f})',
                       fontsize=title_fontsize)
@@ -1333,7 +1451,7 @@ class MafFrame:
             >>> mf.plot_tmb()
             >>> plt.tight_layout()
         """
-        df = self.compute_tmb()
+        df = self.matrix_tmb()
         if samples is not None:
             df = df.loc[samples]
         if ax is None:
@@ -1381,7 +1499,7 @@ class MafFrame:
             >>> mf.plot_vaf('i_TumorVAF_WU')
             >>> plt.tight_layout()
         """
-        genes = self.compute_genes(count=count).index.to_list()
+        genes = self.matrix_genes(count=count).index.to_list()
         s = self.df.groupby('Hugo_Symbol')[col].median()
         genes = s[genes].sort_values(ascending=False).index.to_list()
         df = self.df[self.df.Hugo_Symbol.isin(genes)]
@@ -1470,7 +1588,7 @@ class MafFrame:
             >>> mf.plot_varsum()
             >>> plt.tight_layout()
         """
-        df = self.compute_tmb()
+        df = self.matrix_tmb()
         df = pd.melt(df, value_vars=df.columns)
         if ax is None:
             fig, ax = plt.subplots(figsize=figsize)
@@ -1527,20 +1645,29 @@ class MafFrame:
         ax.set_ylabel('')
         return ax
 
-    def plot_waterfall(self, count=10, ax=None, figsize=None, **kwargs):
-        """Create a waterfall plot.
+    def plot_waterfall(
+        self, count=10, keep_empty=False, samples=None, ax=None, figsize=None, **kwargs
+    ):
+        """
+        Create a waterfall plot.
+
+        See the :ref:`tutorials:Create customized oncoplots` tutorial to
+        learn how to create customized oncoplots.
 
         Parameters
         ----------
         count : int, default: 10
             Number of top mutated genes to display.
+        keep_empty : bool, default: False
+            If True, display samples that do not have any mutations.
+        samples : list, optional
+            List of samples to display.
         ax : matplotlib.axes.Axes, optional
             Pre-existing axes for the plot. Otherwise, crete a new one.
         figsize : tuple, optional
             Width, height in inches. Format: (float, float).
         kwargs
             Other keyword arguments will be passed down to
-            :meth:`matplotlib.axes.Axes.pcolormesh()` and
             :meth:`seaborn.heatmap`.
 
         Returns
@@ -1561,7 +1688,10 @@ class MafFrame:
             >>> mf.plot_waterfall(linewidths=0.5)
             >>> plt.tight_layout()
         """
-        df = self.compute_waterfall(count)
+        df = self.matrix_waterfall(count=count, keep_empty=keep_empty)
+
+        if samples is not None:
+            df = df[samples]
 
         # Apply the mapping between items and integers.
         l = reversed(NONSYN_NAMES + ['Multi_Hit', 'None'])
@@ -1711,8 +1841,14 @@ class MafFrame:
                 return r
             df = df.apply(one_row, axis=1)
 
+        # Create the metadata.
+        meta = [
+            '##fileformat=VCFv4.3',
+            '##source=fuc.api.pymaf.MafFrame.to_vcf',
+        ]
+
         # Create the VcfFrame.
-        vf = pyvcf.VcfFrame(['##fileformat=VCFv4.3'], df)
+        vf = pyvcf.VcfFrame(meta, df)
 
         return vf
 
@@ -1736,3 +1872,38 @@ class MafFrame:
             String representation of MafFrame.
         """
         return self.df.to_csv(index=False, sep='\t')
+
+    def filter_annot(self, af, expr):
+        """
+        Filter the MafFrame using sample annotation data.
+
+        Samples are selected by querying the columns of an AnnFrame with a
+        boolean expression. Samples not present in the MafFrame will be
+        excluded automatically.
+
+        Parameters
+        ----------
+        af : AnnFrame
+            AnnFrame with sample annotaton data.
+        expr : str
+            Query expression to evaluate.
+
+        Returns
+        -------
+        MafFrame
+            Filtered MafFrame.
+
+        Examples
+        --------
+
+        >>> from fuc import common, pymaf
+        >>> common.load_dataset('tcga-laml')
+        >>> mf = pymaf.MafFrame.from_file('~/fuc-data/tcga-laml/tcga_laml.maf.gz')
+        >>> af = pymaf.AnnFrame.from_file('~/fuc-data/tcga-laml/tcga_laml_annot.tsv')
+        >>> filtered_mf = mf.filter_annot(af, "FAB_classification == 'M4'")
+        """
+        samples = af.df.query(expr).index
+        i = self.df.Tumor_Sample_Barcode.isin(samples)
+        df = self.df[i]
+        mf = self.__class__(df)
+        return mf