+++ fuc/cli/bam_index.py
@@ -1,9 +1,9 @@
 from .. import api
+
 import pysam
 
 description = f"""
-This command will index a SAM/BAM/CRAM file. It essentially wraps the
-'pysam.index' method from the pysam package.
+This command will index the input SAM/BAM/CRAM file.
 
 usage examples:
   $ fuc {api.common._script_name()} in.sam
@@ -15,10 +15,10 @@ def create_parser(subparsers):
     parser = api.common._add_parser(
         subparsers,
         api.common._script_name(),
-        help='[BAM] index a SAM/BAM/CRAM file',
+        help='[BAM] Index a SAM/BAM/CRAM file.',
         description=description,
     )
-    parser.add_argument('bam', help='SAM/BAM/CRAM file')
+    parser.add_argument('bam', help='SAM/BAM/CRAM file.')
 
 def main(args):
     pysam.index(args.bam)