+++ fuc/cli/bed_sum.py
@@ -16,7 +16,7 @@ def create_parser(subparsers):
     parser = api.common._add_parser(
         subparsers,
         api.common._script_name(),
-        help='[BED] summarize a BED file',
+        help='[BED] Summarize a BED file.',
         description=description,
     )
     parser.add_argument('bed', help='BED file')