+++ fuc/cli/fuc_demux.py
@@ -23,7 +23,7 @@ def create_parser(subparsers):
     parser = api.common._add_parser(
         subparsers,
         api.common._script_name(),
-        help='[FUC] parse Reports directory from bcl2fastq or bcl2fastq2',
+        help='[FUC] Parse the Reports directory from bcl2fastq.',
         description=description,
     )
     parser.add_argument('reports_dir', help='Reports directory')