+++ fuc/cli/maf_maf2vcf.py
@@ -33,7 +33,7 @@ def create_parser(subparsers):
     parser = api.common._add_parser(
         subparsers,
         api.common._script_name(),
-        help='[MAF] convert a MAF file to a VCF file',
+        help='[MAF] Convert a MAF file to a VCF file.',
         description=description,
     )
     parser.add_argument('maf', help='MAF file')