+++ fuc/cli/tbl_merge.py
@@ -19,7 +19,7 @@ def create_parser(subparsers):
     parser = api.common._add_parser(
         subparsers,
         api.common._script_name(),
-        help='[TABLE] merge two table files',
+        help='[TABLE] Merge two table files.',
         description=description,
     )
     parser.add_argument('left', help='left file')