+++ fuc/cli/vcf_merge.py
@@ -16,7 +16,7 @@ def create_parser(subparsers):
     parser = api.common._add_parser(
         subparsers,
         api.common._script_name(),
-        help='[VCF] merge two or more VCF files',
+        help='[VCF] Merge two or more VCF files.',
         description=description,
     )
     parser.add_argument('vcf_files', help='VCF files', nargs='+')