+++ fuc/cli/vcf_slice.py
@@ -16,7 +16,7 @@ def create_parser(subparsers):
     parser = api.common._add_parser(
         subparsers,
         api.common._script_name(),
-        help='[VCF] slice a VCF file',
+        help='[VCF] Slice a VCF file.',
         description=description,
     )
     parser.add_argument('vcf', help='VCF file')