+++ CHANGELOG.rst
@@ -1,6 +1,22 @@
 Changelog
 *********
 
+0.15.0 (2021-06-24)
+-------------------
+
+* Update :command:`vcf_filter` command.
+* Update :command:`tbl_sum` command.
+* Add ``samples`` and ``shape`` attributes to :class:`pymaf.AnnFrame` class.
+* Rename :meth:`pymaf.MafFrame.compute_genes/tmb/waterfall` methods to :meth:`pymaf.MafFrame.matrix_genes/tmb/waterfall`.
+* Add ``keep_empty`` argument to :meth:`pymaf.MafFrame.matrix_waterfall/plot_oncoplot/plot_waterfall` methods.
+* Add :meth:`pymaf.MafFrame.filter_annot` method.
+* Add :meth:`pymaf.AnnFrame.sorted_samples` method.
+* Fix minor bug in :meth:`pymaf.MafFrame.to_frame` method.
+* Deprecate :meth:`pyvep.filter_lof/clinsig` methods.
+* Update :meth:`pymaf.MafFrame.from_vcf` method to extract genotype keys (e.g. DP, AD, AF).
+* Update :command:`bam_slice` and :command:`bam_rename` commands.
+* Deprecate :meth:`pybam.rename` method.
+
 0.14.0 (2021-06-20)
 -------------------